add ColumnDistributionMetric
