+++ src/evidently/metrics/__init__.py
@@ -9,11 +9,11 @@ from .data_integrity.column_missing_values_metric import ColumnMissingValuesMetr
 from .data_integrity.column_regexp_metric import ColumnRegExpMetric
 from .data_integrity.dataset_missing_values_metric import DatasetMissingValuesMetric
 from .data_integrity_metrics import DataIntegrityMetrics
+from .data_quality.data_quality_value_list_metric import DataQualityValueListMetrics
+from .data_quality.data_quality_value_quantile_metric import DataQualityValueQuantileMetric
 from .data_quality.data_quality_value_range_metric import DataQualityValueRangeMetric
 from .data_quality_metrics import DataQualityCorrelationMetrics
 from .data_quality_metrics import DataQualityMetrics
 from .data_quality_metrics import DataQualityStabilityMetrics
-from .data_quality_metrics import DataQualityValueListMetrics
-from .data_quality_metrics import DataQualityValueQuantileMetric
 from .num_target_drift_metrics import NumTargetDriftMetrics
 from .regression_performance_metrics import RegressionPerformanceMetrics
