+++ tests/calculations/test_data_quality.py
@@ -2,6 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
+from evidently.calculations.data_quality import calculate_column_distribution
 from evidently.calculations.data_quality import get_rows_count
 
 
@@ -17,3 +18,14 @@ from evidently.calculations.data_quality import get_rows_count
 )
 def test_get_rows_count(dataset: pd.DataFrame, expected_rows: int) -> None:
     assert get_rows_count(dataset) == expected_rows
+
+
+@pytest.mark.parametrize(
+    "dataset, bins_count, expected_distribution",
+    (
+        (pd.DataFrame({"test": []}), 3, []),
+        (pd.DataFrame({"test": [1, 2, 1, 2]}), 2, [[1.0, 1.0], [1.0, 1.5, 2.0]]),
+    ),
+)
+def test_calculate_column_distribution(dataset: pd.DataFrame, bins_count: int, expected_distribution: list) -> None:
+    assert calculate_column_distribution(dataset["test"], bins_count=bins_count) == expected_distribution
